+++ app/comments/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CommentsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'app.comments'
