+++ templates/base/edit_profile.html
@@ -1,83 +1,83 @@
-<!doctype html>
-<html lang="es">
-<head>
-    {% load static %}
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Editar Perfil</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="{% static 'css/profile.css' %}">
-</head>
-<body>
+{% extends 'base.html' %}
+{% load static %}
 
-<div class="container-xl px-4 mt-4">
-    <nav class="nav nav-borders">
-        <a class="nav-link active ms-0" href="#">Editar Perfil</a>
-    </nav>
-    <hr class="mt-0 mb-4">
+{% block custom_styles %}
+    <link rel="stylesheet" href="{% static 'css/profile.css' %}">
+{% endblock %}
 
-    {% if messages %}
-        {% for message in messages %}
-            <div class="alert alert-success">{{ message }}</div>
-        {% endfor %}
-    {% endif %}
+{% block container %}
 
-    {% if error %}
-        <div class="alert alert-danger">{{ error }}</div>
-    {% endif %}
-    <form method="POST" enctype="multipart/form-data">
-        {% csrf_token %}
-    <div class="row">
-        <!-- Columna foto de perfil -->
-        <div class="col-xl-4">
-            <div class="card mb-4 mb-xl-0">
-                <div class="card-header">Foto de perfil</div>
-                <div class="card-body text-center">
-                <!-- Imagen de perfil con previsualización -->
-                    <img id="profileImage" class="img-account-profile rounded-circle mb-2"
-                        src="{% if user.profile_photo %}{{ user.profile_photo.url }}{% else %}http://bootdey.com/img/Content/avatar/avatar1.png{% endif %}"
-                        alt="Perfil" style="width:150px; height:150px; object-fit:cover;">
+    <div class="container-xl px-4 mt-4">
+        <nav class="nav nav-borders">
+            <a class="nav-link active ms-0" href="#">Editar Perfil</a>
+        </nav>
+        <hr class="mt-0 mb-4">
 
-                <div class="small font-italic text-muted mb-4">JPG o PNG no mayor a 5 MB</div>
+        {% if messages %}
+            {% for message in messages %}
+                <div class="alert alert-success">{{ message }}</div>
+            {% endfor %}
+        {% endif %}
 
-                <!-- Input oculto para cargar archivo -->
-                <input type="file" id="imageUpload" name="profile_image" accept="image/*" style="display:none;">
+        {% if error %}
+            <div class="alert alert-danger">{{ error }}</div>
+        {% endif %}
+        <form method="POST" enctype="multipart/form-data">
+            {% csrf_token %}
+            <div class="row">
+                <!-- Columna foto de perfil -->
+                <div class="col-xl-4">
+                    <div class="card mb-4 mb-xl-0">
+                        <div class="card-header">Foto de perfil</div>
+                        <div class="card-body text-center">
+                            <!-- Imagen de perfil con previsualización -->
+                            <img id="profileImage" class="img-account-profile rounded-circle mb-2"
+                                 src="{% if user.profile_photo %}{{ user.profile_photo.url }}{% else %}
+                             http://bootdey.com/img/Content/avatar/avatar1.png{% endif %}"
+                                 alt="Perfil" style="width:150px; height:150px; object-fit:cover;">
 
-                <!-- Botón para abrir el selector -->
-                <button class="btn btn-primary" type="button" onclick="document.getElementById('imageUpload').click();">
-                    Subir nueva imagen
-                </button>
-            </div>
-        </div>
-    </div>
+                            <div class="small font-italic text-muted mb-4">JPG o PNG no mayor a 5 MB</div>
 
-        <!-- Columna detalles de cuenta -->
-        <div class="col-xl-8">
-            <div class="card mb-4">
-                <div class="card-header">Detalles de la cuenta</div>
-                <div class="card-body">
-                        <!-- Nombre y Apellido -->
-                        <div class="row gx-3 mb-3">
-                            <div class="col-md-6">
-                                <label class="small mb-1">Nombre</label>
-                                <input class="form-control" type="text" name="first_name" value="{{ user.user.first_name }}">
-                            </div>
-                            <div class="col-md-6">
-                                <label class="small mb-1">Apellido</label>
-                                <input class="form-control" type="text" name="last_name" value="{{ user.user.last_name }}">
-                            </div>
+                            <!-- Input oculto para cargar archivo -->
+                            <input type="file" id="imageUpload" name="profile_image" accept="image/*"
+                                   style="display:none;">
+
+                            <!-- Botón para abrir el selector -->
+                            <button class="btn btn-primary" type="button"
+                                    onclick="document.getElementById('imageUpload').click();">
+                                Subir nueva imagen
+                            </button>
                         </div>
-                        <button class="btn btn-primary" type="submit">Guardar cambios</button>
+                    </div>
+                </div>
+
+                <!-- Columna detalles de cuenta -->
+                <div class="col-xl-8">
+                    <div class="card mb-4">
+                        <div class="card-header">Detalles de la cuenta</div>
+                        <div class="card-body">
+                            <!-- Nombre y Apellido -->
+                            <div class="row gx-3 mb-3">
+                                <div class="col-md-6">
+                                    <label class="small mb-1">Nombre</label>
+                                    <input class="form-control" type="text" name="first_name"
+                                           value="{{ user.user.first_name }}">
+                                </div>
+                                <div class="col-md-6">
+                                    <label class="small mb-1">Apellido</label>
+                                    <input class="form-control" type="text" name="last_name"
+                                           value="{{ user.user.last_name }}">
+                                </div>
+                            </div>
+                            <button class="btn btn-primary" type="submit">Guardar cambios</button>
+                        </div>
+                    </div>
                 </div>
             </div>
-        </div>
+        </form>
     </div>
-    </form>
-</div>
 
-<!-- Bootstrap JS -->
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-<script src="{% static 'js/edit_profile.js' %}"></script>
-
-</body>
-</html>
+{% endblock %}
+{% block js %}
+    <script src="{% static 'js/edit_profile.js' %}"></script>
+{% endblock %}
