+++ templates/school/partial/careers.html
@@ -1,0 +1,1 @@
+<h2>Carreras</h2>
